+++ client/app/modules/login/login.component.spec.js
@@ -3,15 +3,15 @@
 import angular from 'angular';
 import ngResource from 'angular-resource';
 import uiBootstrap from 'angular-ui-bootstrap';
-import login from './login.component';
+import LoginModule from './login.module';
 import AuthModule from '../../modules/auth/auth.module';
 
-describe('Component: LoginComponent', function() {
+describe('Module: LoginComponent', function() {
   // load the component's module
   beforeEach(angular.mock.module(uiBootstrap));
   beforeEach(angular.mock.module(ngResource));
   beforeEach(angular.mock.module(AuthModule));
-  beforeEach(angular.mock.module(login));
+  beforeEach(angular.mock.module(LoginModule));
 
   let loginComponent;